+++ notemaker/src/App.css
@@ -14,7 +14,7 @@
 }
 
 
-@keyframes App-logo-spin {
+@keyframes daisy {
   from {
     transform: rotate(0deg);
   }
